+++ configs/building.txt
@@ -1,0 +1,24 @@
+expname = building
+basedir = ./logs
+datadir = ./data/nerf_synthetic/building
+dataset_type = blender
+training = True
+
+no_batching = True
+
+use_viewdirs = True
+white_bkgd = True
+lrate_decay = 500
+
+N_samples = 64
+N_importance = 128
+N_rand = 1024
+
+precrop_iters = 500
+precrop_frac = 0.5
+
+half_res = True
+
+i_testset = 10000
+
+i_video = 200000
