+++ configs/stonehenge.txt
@@ -1,0 +1,29 @@
+expname = stonehenge_test
+basedir = ./logs
+datadir = ./data/nerf_synthetic/stonehenge
+dataset_type = blender
+
+no_batching = True
+
+use_viewdirs = True
+white_bkgd = True
+lrate_decay = 500
+
+N_samples = 64
+N_importance = 128
+N_rand = 1024
+
+multires = 15
+multires_views = 6
+
+precrop_iters = 500
+precrop_frac = 0.5
+
+half_res = True
+
+i_testset = 10000
+
+i_video = 500000
+
+
+
