+++ configs/teddy_bear.txt
@@ -1,0 +1,23 @@
+expname = teddy_bear_test
+basedir = ./logs
+datadir = ./data/nerf_synthetic/teddy_bear
+dataset_type = blender
+
+no_batching = True
+
+use_viewdirs = True
+white_bkgd = True
+lrate_decay = 500
+
+N_samples = 64
+N_importance = 128
+N_rand = 1024
+
+precrop_iters = 500
+precrop_frac = 0.5
+
+half_res = True
+
+i_testset = 10000
+
+i_video = 200000
